chore: move composable-keys into core
