+++ packages/nuxt/test/composable-keys.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import { detectImportNames } from '../src/plugins/composable-keys'
+import { detectImportNames } from '../../nuxt/src/core/plugins/composable-keys'
 
 describe('detectImportNames', () => {
   const keyedComposables = {
