+++ packages/vite/src/vite.ts
@@ -14,7 +14,7 @@ import { buildClient } from './client'
 import { buildServer } from './server'
 import { warmupViteServer } from './utils/warmup'
 import { resolveCSSOptions } from './css'
-import { composableKeysPlugin } from './plugins/composable-keys'
+import { composableKeysPlugin } from '../../nuxt/src/core/plugins/composable-keys'
 import { logLevelMap } from './utils/logger'
 import { ssrStylesPlugin } from './plugins/ssr-styles'
 import { VitePublicDirsPlugin } from './plugins/public-dirs'
